efm_service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
groq==0.5.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
numpy==1.26.4
cornac==2.2.1
pandas==2.1.4
#Python 3.12